Fix flaky main.range_icp test
Summary: The main.range_icp test failed with a different execution plan for one of the queries in main.range_icp (t10182158).  I could not reproduce the problem even when crushing the test.  Since the most common reason to produce a different execution plan would be different statistics and since I don't know if statistics in InnoDB are deterministic or not, I am adding a call to ANALYZE TABLE before querying the execution plan in order to make sure the statistics are the same values every time.

Test Plan: MTR

Reviewers: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D55353