Per-user session variable default values
Summary:
This diff introduces per-user session variable default values. This is for MyRocks and MySQL production debugging. A new global variable 'per-user-session-var-default-val' with type of string is introduced, it defines the default values of multiple session variables for multiple users, e.g.

--per-user-session-var-default-val="user1:user2:user3:tx_isolation=read-committed,gap_lock_raise_error=0,user4:user5:big_tables=1,auto_increment_increment=12345,long_query_time=123.45"

In this example, the default values of session variables tx_isolation and gap_lock_raise_error for user1, user2, and user3 are defined, the default values of session variables big_tables, auto_increment_increment, and long_query_time for user4 and user 5 are defined. When these users connect to server, the corresponding session variables will be set with these values so that these users don't have to execute SET command to set these session variables before executing queries. The cost of an extra round trip is saved. This global variable is dynamic so it can be set at runtime. The supported types are enum/bool/int/uint/double, which are the most common types for session variables.

Test Plan:
New tests added:
  sys_vars.per_user_session_var_default_val_basic.test
  sys_vars.per_user_session_var_default_val_null.test
  sys_vars.per_user_session_var_default_val_empty.test
  sys_vars.per_user_session_var_default_val.test

Reviewers: santoshb, jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D55791