Avoid reallocating transaction objects
Summary: The BeginTransaction function in RocksDB was recently updated to allow passing an old object for reuse to avoid reallacating these objects all the time.  This diff is to update MyRocks to take advantage of this feature.

Test Plan: MTR - and rand Valgrind on a large portion of the tests

Reviewers: hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D55857