Enable external TSan build support on Ubuntu 15.10
Summary: This diff enables TSan build support on Ubuntu 15.10. It will most likely work on older versions of Ubuntu as well, but I haven't tested on them, so being very explicit here. It is unknown if this solution works on Fedora 23 or not. Would appreciate someone trying it out.

Test Plan:
  - Build with `-DWITH_TSAN=1`.
  - Make sure that resulting build succeeds, `mysqld` starts up and displays warnings from TSan.

Reviewers: jtolmer

Reviewed By: jtolmer

Subscribers: MarkCallaghan, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D56289