Wait properly for RocksDB threads to finish before unloading MyRocks plug-in
Summary:
Current solution doesn't wait till the thread has actually finished. It waits when mutex inside the thread function is released. Though mutex is released at the end of the function this doesn't in any way guarantee that thread has completed, resources associated with the thread have been freed etc. This type of pattern is tends to cause bugs which are hard to debug and depend on timing and scheduling.

Correct way to is to use `pthread_join` to wait till all the resources associated with the thread are cleaned up and the thread is really done.

Reviewed By: vasilep

Differential Revision: https://reviews.facebook.net/D56205

fbshipit-source-id: ccea0e4a726