Add row information to STATUS logs
Summary:
The audit plugin requires that row information be included in the STATUS log lines. Previously, they were only present in the RESULT logs, but we wanted to merge STATUS and RESULT so that we're not emitting too much redundant information.

Squash with: https://reviews.facebook.net/D49329

Test Plan:
mtr

Also, tested with audit plugin.

Reviewers: jkedgar

Reviewed By: jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D58419