Make the server error ER_CONNECTION_TIMEOUT error number generated in the default way
Summary:
In https://reviews.facebook.net/D54213, to make the new ER_CONNECTION_TIMEOUT "Connection closed due to timeout." be compatible with the client error 2006, the error code was made to 2006 by adding place holders between the current last error 1921 at that time and error 2005, actually this is not good since error code starting with 2000 in MySQL are used by client, and it is not necessary for the new feature. Server shouldn't return error code greater than 1999. In this diff, all the place holders are removed, when pushing the connection timeout error into socket, error code 2006 is used directly and the actual error number of ER_CONNECTION_TIMEOUT is ignored.

Squash with: e04c9e1  https://reviews.facebook.net/D54213

Test Plan: MTR

Reviewers: santoshb

Subscribers: webscalesql-eng