Don't restart mysqld if crashes happen during shutdown
Summary: We saw copies failed when mysqld crashes during shutdown in prod. The issue here is that when crashes happen during shutdown, mysqld_safe will restart mysqld, which fails the original shutdown. To fix this issue, this diff introduces shutdown file. This is an empty file with name in the form of <pid file name>.shutdown, i.e. pid file name with a suffix of '.shutdown', mysqld will create this file when a regular shutdown starts, and mysqld_safe won't restart mysqld if this file presents. Both mysqld_safe and mysqld will delete this file at start time if it exists.

Test Plan:
-- New MTR test mysqld_shutdown_file.test
-- Manually tested in production environment

Reviewers: santoshb, jkedgar

Subscribers: webscalesql-eng