Change handshake to send server version, but append MYSQL_COMPILATION_COMMENT at the end.
Summary:
This is needed because the change to use comment inside the handshake caused
the JDBC connector to break.  Several of the clients expect the server to pass
the version in X.XX.XX format.

https://github.com/facebook/mysql-5.6/issues/233#issuecomment-214439969 for
more details.

Squash With D56547.

Test Plan: Verify again using `ngrep`, verified linkbench error went away on local machine for JDBC.

Reviewers: jtolmer, anca

Reviewed By: anca

Subscribers: santoshb, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D57183