Fixed innodb.innodb-status-output
Summary:
The innodb.innodb-status-output test (which only runs when --big-test is specified) depended on three lines being adjacent in the error log.  A recent change to add git hash/date into the log put the new lines in between two of those three lines causing this test to fail (t11045584).  This diff moves the two lines prior to the first of the three lines that the test depends on.

Squash with: https://reviews.facebook.net/D56355

Test Plan: MTR and 'mysqltest.sh --big-test innodb.innodb-status-output'

Reviewers: hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D57429