Fix git hashes and dates when building RPMs
Summary: The RPM build script does things in such a way that the current method of getting the git hashes and dates for the main repository and rocksdb submodule fails.  Update the CMake files and perl scripts so the git hashes and dates can be passed in when using mysqlrpm.sh.

Test Plan: Build an RPM using mysqlrpm.sh and verify that the git hashes and dates are correct.

Reviewers: mung, hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D57579