change check_if_supported_inplace_alter to use guard clause to prevent inplace alter on operations that are not yet supported
Summary:
When dropping a column on a table, if that column is also part of an SK, the
inplace drop alter code is triggered to drop the SK.  The inplace alter code
does not have the implementation to handle drop column yet.

Similarly when executing a statement which both DROPS and ADDS an index in the same statement, an online alter would be attempted, but the ADD index functionality is not supported yet, causing assertion fail in MyRocks.

Workaround is to only do inplace alter when the operation is a DROP INDEX on secondary key by itself, and not combined with other operations which arent supported yet.

Fixes https://github.com/facebook/mysql-5.6/issues/239

Squash with D54777

Test Plan: Added new test cases to drop_index_inplace that reproduces the issue

Reviewers: hermanlee4, yoshinorim

Reviewed By: yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D57585