Fix a corruption error related to NULL columns.
Summary:
The MyRocks code has some field_decoders structures that allow columns that are not needed to not be decoded when converting from the on-disk data format to the MySQL in-memory format.  These were not set up correctly when nullable columns were part of the row and would cause us to attempt to skip too many bytes in the row - causing either incorrect data or an error indicating corruption.  (Issue #238)

Squash with: https://reviews.facebook.net/D56115

Test Plan: Updated rocksdb.col_opt_null to add a query that fails with the old code and verified that it succeeded with the new code.

Reviewers: hermanlee4, yoshinorim, spetrunia

Reviewed By: spetrunia

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D57633