Fix an error indicating corruption that occurs when selecting specific columns from a table when other columns prior to the selected columns (which are not selected and are not a part of the primary key) are NULL.
Summary: The MyRocks code has some field_decoders structures that allow columns that are not needed to not be decoded when converting from the on-disk data format to the MySQL in-memory format.  These were not set up correctly when nullable columns were part of the row and would cause us to attempt to skip too many bytes in the row - causing either incorrect data or an error indicating corruption.  (Issue #238)

Test Plan: Updated rocksdb.col_opt_null to add a query that fails with the old code and verified that it succeeded with the new code.

Reviewers: yoshinorim, hermanlee4

Subscribers: