Support index-only scans for simple collations
Summary:
For binary collations, index-only scans are supported by decoding the memcomparable form back into the original form. This is because the Field::make_sort_key function is invertible for binary collations.

This adds support for index-only scans for 'simple' collations. Simple collations map a single character to a single character. If they map to the same character, then we lose information about what the original character was.

The solution is for every character with such conflicts, we add an index to unpack_info specifying which source character was used to create the sort key. For example, if 'a', 'A' and 'à' all map to 'A', then they would get assigned indices 0, 1 and 2 respectively.

If a character has no conflict, nothing is stored in unpack_info. For example, `As.Df` converts to `AS.DF`. The unpack info stored would be `0101`. Nothing is stored for '.' because '.' maps uniquely to '.'.

Linkbench shows that having minimal unpack data gives better performance:
Minimal unpack info:
  table_name      table_rows      avg_row_length  data_length     max_data_length index_length
  linktable       4375367708      19      84724839472     0       113804161473
  INFO 2016-05-12 15:03:45,930 [main]: REQUEST PHASE COMPLETED. 50000000 requests done in 1015 seconds. Requests/second = 49244

Complete unpack info:
  table_name      table_rows      avg_row_length  data_length     max_data_length index_length
  linktable       4371383928      19      85532874063     0       128972663377
  INFO 2016-05-12 15:04:17,576 [main]: REQUEST PHASE COMPLETED. 50000000 requests done in 1045 seconds. Requests/second = 47818

No unpack info:
  table_name      table_rows      avg_row_length  data_length     max_data_length index_length
  linktable       4371383928      19      85229266495     0       100223719092
  INFO 2016-05-11 12:06:19,509 [main]: REQUEST PHASE COMPLETED. 50000000 requests done in 2982 seconds. Requests/second = 16764

Note that this only adds support for using unpack_info for secondary keys. For primary keys, we still store the full original string in the value portion, and the sort key in the key portion.

Also note that previously existing secondary indices should work just fine with the new code. If we find a row where unpack_info is needed, but there is none provided, then we will just do a PK lookup. The explain output will still show that we're doing an index-only scan though since you won't know if it's available until you start looking at rows.

In theory, this could be optimized by checking the version number during the optimization phase, but ha_rocksdb::index_flags may be called before ha_rocksdb::open is called, which means that we don't always have access to this information.

Test Plan:
rocksdb.type_char_indexes_collation

Manually set version to SECONDARY_FORMAT_VERSION_UNPACK_INFO, and then ran:
  mysqltest.sh --testset=RocksDB

Reviewers: spetrunia, hermanlee4, jkedgar

Reviewed By: jkedgar

Subscribers: vasilep, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D58269
Differential Revision: https://reviews.facebook.net/D62121
Differential Revision: https://reviews.facebook.net/D62793