Issue #261: Fix blob/text memcomparable length
Summary:
Blob and text columns were not getting the appropriate key length because it was using Field::pack_length() to get the key length. However, if you look at Field_blob::pack_length, the value returned there is # bytes used for length + sizeof(char*). This is not valid.

Instead we should be grabbing the key prefix size from KEY_PART_INFO. This is passed down to Rdb_field_packing::setup for use.

Differential Revision: https://reviews.facebook.net/D59439

fbshipit-source-id: d826b790a47