Fix sort order for empty strings in 'text' columns
Summary:
This is a fix for an upstream bug: http://bugs.mysql.com/bug.php?id=81810

The problem is that Field_blob::make_sort_key will memzero the sort key for the empty string. However, for most collations, the empty string should sort as spaces, not as NUL characters.

The fix is to memzero only if the collation pad character happens to be 0 as well.

Test Plan: mtr

Reviewers: jkedgar, hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D59595