Fix length in Field_blob::make_sort_key
Summary:
For the binary collation in Field_blob::make_sort_key, we write out the length of the entire blob, even though the key itself may be shorter. We can miss out on duplicate key errors because of this:

Suppose key length is 1, and we want to encode "ab", "a". Then we get the following sort keys:

  'a' 0 2
  'a' 0 1

We want the two of them to produce the same key instead. To fix, define key length to be the min of blob length and requested key length.

Test Plan: mtr

Reviewers: jkedgar, hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D59631