Issue #266: Fix varchar length unpacking for updates
Summary:
For varchars with collations that support unpacking, we are unpacking the length of the varchar into the Field->ptr buffer.

However, this is wrong because the actual buffer may have a row_offset, which should be determined by adding ptr_diff.

The fix is to unpack the length into the given destination buffer, which already points to the buffer with the correct offset.

Fixes #266

Differential Revision: https://reviews.facebook.net/D59685