Issue #259: Fix unintialized memory in CHAR records
Summary:
For the collations that we currently support, mbmaxlen is always 1 except for utf8_bin.

For utf8_bin, mbmaxlen is 3, which means that a character can take up to 3 bytes. When allocating space for it, we use Field_string::pack_length() = character length * mbmaxlen.

However, not all utf8 characters will take up all 3 bytes, so the actual string may be less than Field_string::pack_length() In SQL, this is usually padded with spaces. When unpacking in MyRocks, if we don't have enough characters to fill up pack_length(), we should be filling it with spaces instead of leaving it uninitialized.

VARCHARs are immune because they have a length associated with them. For CHARs, there isn't a length, so the whole field is expected to contain valid data.

Fixes #259.

Reviewed By: jkedgar

Differential Revision: https://reviews.facebook.net/D59727