Update RocksDB submodule snapshot and fix all the associated issues
Summary:
Continuous integrations shows that recent changes in RocksDB are causing MySQL builds to fail. Build breaks are caused by renaming `memtable_prefix_bloom_bits` to `memtable_prefix_bloom_probes`. In addition we need to remove the usage of `memtable_prefix_bloom_probes` because that option is deprecated.

When it comes to test failures then a change to the `PerfLevel` enumeration value start index is causing a number of breaks as well. That's because we hard-code the enumeration values in our tests. There's no good way to fix this easily.

This also proved to be a good opportunity to simplify the code dealing with `perf_context_level` - I added stronger validation as part of `MYSQL_THDVAR_INT` macro which exposes this value. This way we can internally assume valid ranges.

update-submodule: rocksdb

Differential Revision: https://reviews.facebook.net/D59721

fbshipit-source-id: e16100035d4