Fix sbm during restarts
Summary:
SHOW SLAVE STATUS returns a value of 0 for seconds_behind_master(sbm) during restarts. This is invalid as
the real lag is only possible to know only after sql thread receives and executes binlog events. The invalid sbm is
due to the fact that slave_has_caughtup is initialized to True. Fix this by maintaining three states for the variable
slave_has_caughtup (i,e; NONE, YES, NO). The value of sbm will be NULL, 0 and real lag for the respective states.

Test Plan: Added a mtr test

Reviewers: pengt, tengli

Reviewed By: tengli

Subscribers: webscalesql-eng, ebergen

Differential Revision: https://reviews.facebook.net/D60249