More OSS goodness: fix compilation errors with Clang 3.8 on Ubuntu 16.04 LTS
Summary:
Clang is smart enough to understand that pointer is being compared to `'\0'` which is probably not the intent of the comparison. This causes a warning to be issued which in turn triggers a build break because we treat warnings as errors.

The main reason for fixing this is to make sure that we can cleanly compile our branch on Ubuntu 16.04 LTS with Clang 3.8.

Test Plan: Follow the instructions at https://github.com/facebook/mysql-5.6/wiki/Build-Steps and verify that after applying the fixes we'll have a clean build.

Reviewers: mung

Reviewed By: mung

Subscribers: webscalesql-eng, MarkCallaghan

Differential Revision: https://reviews.facebook.net/D60795