Add current master host:port to server read only error when available
Summary:
When a server is in read_only, we append the current master_host and
master_port to the error message, if the server is a slave. The master_host
could be either hostname or ip address, depending on how a slave is connected
to the master.

A new generic error code `ER_OPTION_PREVENTS_STATEMENT_EXTRA_INFO` is added, so
that we can append any additional info to the fixed error message.

Test Plan:
Updated rpl.rpl_read_only. Master_port is non-deterministic in the test, so I
masked in the test result.

Will update any test that comes up in the sandcastle if needed.

Reviewers: santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng, ebergen

Differential Revision: https://reviews.facebook.net/D61383