Add sstfilewriter api improvements to fast secondary index creation (#373)
Summary:
This diff utilizes the recently implemented SSTFileWriter API for fast secondary index creation.

Writing to SST files directly requires the records of each secondary key to be written in order, to do this we need to leverage an external sort as the index is often larger than what can be held memory.

We iterate through the primary key entries, writing the offsets of each entry into a balanced binary tree (std::set, which uses red-black tree under the hood).  The tree itself uses a custom comparator which does a byte-wise comparison of records during insertion, which it uses to maintain the ordering of the tree.

Once we have reached the limit of what our sort buffer can hold in memory (RDB_MERGE_BUF_SIZE), we write to a tmp file out to disk in sorted order by iterating through the sorted "offset tree".  We continue to do this until all entries in the primary key have been traversed.  At this point we have several sorted chunks in a tmp_file on disk.

We now perform the merge portion of the external sort, merging each of the sorted chunks.  We do this by maintaining a heap  (std::priority_queue, again using the custom comparator) which points to the entries in each chunk as we read them off the disk.  Finally, as we pop off each entry from the top of the heap,  we insert it into the SSTFileWriter via the new API.

Fixes https://github.com/facebook/mysql-5.6/issues/277

Reviewed By: jkedgar

Differential Revision: D4102146 (https://github.com/facebook/mysql-5.6/commit/64db3fb336b626d0aa67654553eb86c00366bcb4)

fbshipit-source-id: c89d102e7b7