Add a variable for appending a custom message to server read_only error
Summary:
Proposing a way to set a custom message in the `--read-only (super)` error.
Note: the variable can only be set when server is in read-only (super) state.

p.s. I am hesitant to add a mutex for the variable, so we will get a snapshot
of the current pointer of the variable (const char*) when printing the error.

Squash with https://reviews.facebook.net/D61383

Test Plan: Added tests.

Reviewers: santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng, ebergen

Differential Revision: https://reviews.facebook.net/D61713