Added table checksums for mysqldump with --enable-checksum-table option
Summary:
This diff added an additional option: --enable-checksum-table  that enables the checksums of each table to be printed to the dump as comments.

The checksum of each table is calculated with the actual logical row data in the string format and can be compared against each dump file's table checksum comment to verify the dump's consistency. Later either MySQL infra or Eng can simply write a simple comparison script that compares the checksums of tables between two dumps.

Note that --enable-checksum-table cannot be used with --tab option since this '--tab' essentially does a query of  'SELECT INTO OUTFILE' and there is no way for the mysqldump to get the actual row data. After speaking with Karol and Junyi, we decide to turn this feature off when --tab is enabled and eventually turn off '--tab' for places that uses '--tab'.

Test Plan: mysqltest.sh mysqldump-checksum

Reviewers: santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng, div, karol, junyilu

Differential Revision: https://reviews.facebook.net/D61809
Differential Revision: https://reviews.facebook.net/D65109