Centralize parsing of database/table/partition information
Summary: The Rdb_tbl_def used to keep around just the fully qualified table name which included the database name, the table name, and possibly partition information.  When just one of these was needed this information was parsed but not cached anywhere.  This means that we often parsed this information many multiple times.  Change the code to parse it once and then store it in the Rdb_tbl_def to allow it to be accessed as many times as necessary without further parsing.

Differential Revision: https://reviews.facebook.net/D62073

fbshipit-source-id: febe6079e90