Fix problem with bulk load when a table (ha_rocksdb) is closed early
Summary:
The original code was keeping a list of ha_rocksdb objects that used bulk load in the transaction, but these could be destructed before the transaction was closed.  Add code to close out bulk loading and remove the ha_rocksdb object from the transaction's list when this happens.

Squash with: https://reviews.facebook.net/D59643

Test Plan: MTR

Reviewers: mung, hermanlee4, yoshinorim

Reviewed By: yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D62109