[libmysql] Read compressed packets of size 0xffffff
Summary:
This fixes: http://bugs.mysql.com/bug.php?id=82792

Currently when using the compressed protocol we buffer
received packets and return them as they are requested. We save
the first char of the next packet when returning a packet, and replace
it with a null byte. When we receive large packets (>0xffffff) they
are split over multiple packets. A zero size packet is used to terminate
a multi-packet if it is an exact multiple of 0xffffff. In this case the
character that is saved is the first byte of the 0 byte header, rather
than the first byte of the header of the next packet. When this character
is restored it can result in the next packet being mangled. (size will be
wrong but the sequence number will be fine)

Test Plan: unit tests

Reviewers: mung

Reviewed By: mung

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D62847