Enable mysqlbinlog --skip-empty-trans to handle BEGIN event with empty database in RBR
Summary:
In RBR, it's expected the slave's binlog does not carry a session database name for each BEGIN query event.

However, this affects the functionality of --skip-empty-trans since this feature was designed to assume that each BEGIN query event should always carries a session database.

This diff fixed this issue by relaxing the requirement so that the transaction's database name is set to be the next EVENT 's db name (that carries the database name) after the BEGIN event if the BEGIN doesn't carry the session database name.

Also added some more checking and debugging info to be displayed

MTR test will fail before the change as the reproducing case.

Reference Patch: https://github.com/facebook/mysql-5.6/commit/3c7462a

Reviewed By: abhinav04sharma

Differential Revision: D10115644

fbshipit-source-id: f95b5e13e83