Enable mysqlbinlog --skip-empty-trans to handle BEGIN event with empty database in RBR
Summary:
In RBR, it's expected the slave's binlog does not carry a session database name for each BEGIN query event.

However, this affects the functionality of --skip-empty-trans since this feature was designed to assume that each BEGIN query event should always carries a session database.

This diff fixed this issue by relaxing the requirement so that the transaction's database name is set to be the next EVENT 's db name (that carries the database name) after the BEGIN event if the BEGIN doesn't carry the session database name.

Also added some more checking and debugging info to be displayed

MTR test will fail before the change as the reproducing case.

Test Plan:
MTR test

```
mysqltest.sh rpl.rpl_mysqlbinlog_skip_empty_trans

```
and also tested with a timelinedb's slave binlog on my devserver

```
./_build-5.6-Debug/client/mysqlbinlog --database db171089 --skip-empty-trans --skip-gtids /data/users/tengli/skip-empty-trans-debug/binary-logs-3309.310810.slave
```

Both the MTR and the manual test won't pass before the change

Reviewers: santoshb

Reviewed By: santoshb

Subscribers: vasilep, shlomoid, dreif, div, junyilu, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D63003