[libmysql] Enable Compression Protocol for Async Client
Summary:
Allow the async client to use the compressed protocol.
Most of the structures are logically similar to the sync client,
the main differences revolve around allowing the functions to
be reentered. We maintain Chip's optimisation to minimise extra
copies of the incoming packet but since compression does extra
allocations anyway, we need to manage those extra buffers ourself

Test Plan: Submitted diff which forces compress flag for all tests

Reviewers: mung, anca

Reviewed By: anca

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D62439