Fix MySQL builds with the latest RocksDB
Summary:
Recently RockDB removed a number of options related to compaction (`expanded_compaction_factor`, `source_compaction_factor`, `max_grandparent_overlap_factor`) and replaced them with `max_compaction_bytes`. Currently our validation builds with the latest version of RocksDB are broken because of that. The proper solution would be to adjust the code and ingest the latest version of RocksDB. However ... for various reasons at this point we can't. Therefore we'll implement an ugly workaround which consists of just no longer referencing the obsolete options.

Once we'll move to an updated version of RocksDB then we'll need to implement the usage of `max_compaction_bytes`.

Differential Revision: https://reviews.facebook.net/D63891

fbshipit-source-id: b829cbed3b7