Initialize index stats structure for each key during analyze
Summary:
For tables that have no data, because the stats structure was not
initialized during the analyze scan, the stats data for the table's keys
were never stored and updated in the cached data. This resulted in stale
data remaining behind.

Test Plan:
run unit tests. Tested on udb7838.lla1:3307 and verified the -1 stats
was fixed correctly.

Reviewers: jkedgar, yoshinorim

Reviewed By: yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D64281