Fix Async Compressed protocol
Summary:
When implementing the compressed protocol, I didn't account
for specific orderings of async events. Currently we can clobber a new
packet with an old saved char if the packet is split over multiple network
packets. Modified it to save the char when exitting the function due to
ASYNC_NOT_READY.
It'd be nicer to know when not to use the saved char but I can't think
of a way to do this right now

Rough order of events
Char is saved (doesn't really matter when)
First chunk of packet is present in buffer
When performing read socket is not ready
Loop is exited, but when we return we clobber
the current packet with stale data

squash with D62439

Test Plan: rebuild and repro, passes

Reviewers: mung, anca

Reviewed By: anca

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D64407