Attempt to get more information on rocksdb.rocksdb failure
Summary: I recently updated the timeout on the rocksdb_lock_wait_timeout test from 1000 to 2000 (thinking it was in milliseconds).  Looking closer, it is in seconds, so 1000 should be way more than enough.  I'm not sure what is going on here, but I've modified the test to print out the time it took if it is over the limit to see if we are at the 5000 number or not.  I also added a comment so the next person that runs into this will notify me.

Reviewed By: tianx

Differential Revision: https://reviews.facebook.net/D64437

fbshipit-source-id: 1af8d0b678d